backend_rag/requirements.txt
langchain
chromadb
openai
python-dotenv
langchain-community
langchain-openai
flask
flask-cors
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pypdf